Chapter start and end times are floats; however, some that use only whole seconds do not provide the fraction part. Fix issue with casting int and double.

+++ lib/src/model/chapter.dart
@@ -3,8 +3,8 @@ class Chapter {
   final String imageUrl;
   final String url;
   final bool toc;
-  final int startTime;
-  final int endTime;
+  final double startTime;
+  final double endTime;
 
   Chapter({
     this.title,
